Prevent search/replace controls to submit the container form
The search/replace buttons are seen as "submit" type because the type attribute
is not specified. If ACE is embedded within a form (between <form>...</form>)
clicking on one of these button will submit the outer form.

Setting search/replace buttons type to "button" prevents the outer form to be
submitted.