Merge modelist branch & use it
This branch now depends on the modelist being pulled in as a module. I'm
using the modelist to generate the menu options for switching modes
because it's redundant to have the same massive list in two places.

The modelist module also has some awesome features so I choose it over
my silly list. :D