kitchen_sink & mode_creator using modelist extension
I refactored all requires to point at `ace/ext/modelist` deleted the
build directory and did a full rebuild. In both the kitchen sink and the
mode creator I was able to use the mode lists and see an immediate
reaction in the applications. I think it worked out. :D