Redraw lines and active line highlight when offscreen updates are made
When setUseWrapMode is true, if some text is inserted into a line which is
above the cursor and offscreen, then the renderer did not previously redraw the
lines or active line highlight. However, if this insert causes the line to wrap
onto another line then everything is shifted down by one visual line, leaving
the onscreen lines in an out of date state. With this commit, the onscreen
lines and active line highlight are now redrawn when this happens.