mode: add basic rst syntax highlighting
Add a basic set of tokenizer rules to highlight reStructuredText.

It supports most rst constructs:

* directives
* code-blocks
* lists (bullet and numbered)
* tables
* inline markup (escaped special characters are supported)
* comments

Nested inline markup is not supported (this is not supported by rst
either so that is not a problem).

There is a limitation with markup constructs that span on multiple
lines. ACE tokenizer works line by line. This is not possible to match
bold or italic inline markup that starts on a line and ends on the next
one. Same problem with headings, link references, inline code, etc.

For most tokens, the standard CSS classes are used but these are
special:

table grid characters                 -> ".ace_markup.ace_table"
raw code blocks and directive bodies  -> ".ace_markup.ace_raw"
link definitions and references       -> ".ace_markup.ace_link"
(I did not use markup.underline.link because people may not want links
to be underlined).

Most themes do not define any style for these tokens, a future commit
may need to define style for them.

Signed-off-by: Robin Jarry <robin@jarry.cc>