fix: swap blue/red in cloudeditor themes to reduces usage of red (#5492)
Some users reported that the usage of red for variables could be confusing because users associate red with errors in their code. This swaps blue and red in the themes to reduce the usage of red. As a follow-up we could reduces the bright red-ness of our red to reduce the association with errors.